requirements.txt
asgiref==3.4.0
autopep8==1.5.7
Django==3.2.4
django-bootstrap-form==3.4
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-js-asset==1.2.2
django-subcommands==2020.12.3
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
